+++ Entity/Addon.php
@@ -1,8 +1,16 @@
 <?php
+/*
+ * This file is part of the Sulu CMS.
+ *
+ * (c) MASSIVE ART WebServices GmbH
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
 
 namespace Sulu\Bundle\ProductBundle\Entity;
 
-use Doctrine\ORM\Mapping as ORM;
+use Doctrine\Common\Collections\ArrayCollection;
 
 /**
  * Addon
@@ -10,53 +18,34 @@ use Doctrine\ORM\Mapping as ORM;
 class Addon
 {
     /**
-     * @var string
-     */
-    private $price;
-
-    /**
-     * @var integer
+     * @var int
      */
     private $id;
 
     /**
-     * @var \Sulu\Bundle\ProductBundle\Entity\ProductInterface
+     * @var ProductInterface
      */
     private $product;
 
     /**
-     * @var \Sulu\Bundle\ProductBundle\Entity\ProductInterface
+     * @var ProductInterface
      */
     private $addon;
 
-
     /**
-     * Set price
-     *
-     * @param string $price
-     * @return Addon
+     * @var ArrayCollection
      */
-    public function setPrice($price)
-    {
-        $this->price = $price;
-    
-        return $this;
-    }
+    private $addonPrices;
 
-    /**
-     * Get price
-     *
-     * @return string 
-     */
-    public function getPrice()
+    public function __construct()
     {
-        return $this->price;
+        $this->addonPrices = new ArrayCollection();
     }
 
     /**
      * Get id
      *
-     * @return integer 
+     * @return int
      */
     public function getId()
     {
@@ -66,10 +55,11 @@ class Addon
     /**
      * Set product
      *
-     * @param \Sulu\Bundle\ProductBundle\Entity\ProductInterface $product
-     * @return Addon
+     * @param ProductInterface $product
+     *
+     * @return self
      */
-    public function setProduct(\Sulu\Bundle\ProductBundle\Entity\ProductInterface $product)
+    public function setProduct(ProductInterface $product)
     {
         $this->product = $product;
     
@@ -79,7 +69,7 @@ class Addon
     /**
      * Get product
      *
-     * @return \Sulu\Bundle\ProductBundle\Entity\ProductInterface
+     * @return ProductInterface
      */
     public function getProduct()
     {
@@ -89,10 +79,11 @@ class Addon
     /**
      * Set addon
      *
-     * @param \Sulu\Bundle\ProductBundle\Entity\ProductInterface $addon
-     * @return Addon
+     * @param ProductInterface $addon
+     *
+     * @return self
      */
-    public function setAddon(\Sulu\Bundle\ProductBundle\Entity\ProductInterface $addon)
+    public function setAddon(ProductInterface $addon)
     {
         $this->addon = $addon;
     
@@ -102,10 +93,42 @@ class Addon
     /**
      * Get addon
      *
-     * @return \Sulu\Bundle\ProductBundle\Entity\ProductInterface
+     * @return ProductInterface
      */
     public function getAddon()
     {
         return $this->addon;
+    }
+
+    /**
+     * @return ArrayCollection
+     */
+    public function getAddonPrices()
+    {
+        return $this->addonPrices;
+    }
+
+    /**
+     * @param AddonPrice $addonPrice
+     *
+     * @return self
+     */
+    public function addAddonPrice(AddonPrice $addonPrice)
+    {
+        $this->addonPrices->add($addonPrice);
+
+        return $this;
+    }
+
+    /**
+     * @param AddonPrice $addonPrice
+     *
+     * @return self
+     */
+    public function removeAddonPrice(AddonPrice $addonPrice)
+    {
+        $this->addonPrices->removeElement($addonPrice);
+
+        return $this;
     }
 }